% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_time.R
\name{aggregate_time}
\alias{aggregate_time}
\title{Aggregate time granularity of a data set}
\usage{
aggregate_time(
  x,
  time = NULL,
  grouping_vars = NULL,
  granularity = c("year", "half year", "quarter", "month", "week", "day"),
  adhoc_granularity = NULL,
  aggregation = c("mean", "sum", "max", "min", "median", "count"),
  fill_time_series = FALSE
)
}
\arguments{
\item{x}{Data frame in which observations are labeled with a time or date.}

\item{time}{A string identifying the variable in the data, \code{x},
containing the date/time of observations. Time variable must be one of:
\itemize{
\item \code{Date} object (date or datetime)
\item \code{character} object in ISO calendar month basic format
(e.g. 2022-05)
\item \code{character} object in ISO calendar week extended format
(e.g. 2022-W25)
\item \code{character} object with years and quarters (i.e. 2022-Q2)
}}

\item{grouping_vars}{A string or vector of strings identifying the variables
in the data, \code{x}, containing groups by which time should be aggregated.

If \code{NULL}, data are aggregated without regard to any subgroups in the
data.}

\item{granularity}{Target time granularity for aggregated data set. One of:
\itemize{
\item \code{year}
\item \code{half year}
\item \code{quarter}
\item \code{month}
\item \code{week}
\item \code{day}
}
Ignored if \code{adhoc_granularity} is provided. Note that the requested
granularity must be coarser than the time granularity of the data (\code{x}).
For example, if observations in the data measured at the level of months,
\code{granularity} must be \code{quarter} or \code{year}.}

\item{adhoc_granularity}{A vector of time points defining the ad-hoc time
granularity desired for the aggregation. Provide only the unique time points
for aggregation. Accepted formats are:
\itemize{
\item \code{Date} object (date or datetime)
\item \code{character} object in ISO calendar month basic format
(e.g. 2022-05)
\item \code{character} object in ISO calendar week extended format
(e.g. 2022-W25)
\item \code{numeric} four-digit year
}
Observations are aggregated to the nearest past time point. For example, if
the adhoc granularity includes \code{c(1999, 2004, 2012)}, the years 1999,
2000, 2001, 2002, and 2003 are aggregated to 1999, the years 2004-2011
are aggregated to 2004, and observations at or after 2012 are aggregated to
2012.}

\item{aggregation}{Aggregation function applied when collapsing observations
to the time granularity specified in \code{granularity}. One of:
\itemize{
\item \code{mean} - take mean of all numeric variables by group and time
\item \code{sum} - sum all numeric variables by group and time
\item \code{max} - find maximum of each numeric variable by group and time
\item \code{min} - find minimum of each numeric variable by group and time
\item \code{median} - find median of each numeric variable by group and
time
\item \code{count} - count the number of observations by group and time
}}

\item{fill_time_series}{Logical indicating whether to fill NAs in the output
data for any time points missing from the input data. Ignored if
\code{adhoc_granularity} is provided.}
}
\value{
A tibble, \code{x}, transformed to aggregate it to the desired time
granularity by applying the function specified in \code{aggregation}.
\subsection{ISO 8601 formats}{

Depending on the time granularity chosen, \code{aggregate_time} returns the
appropriately formatted time in a variable titled "jnry_(time)", i.e.:

\itemize{
\item \strong{year} - \code{jnry_year} - ISO calendar year basic format
(e.g. 2022)
\item \strong{half year} - \code{jnry_halfyr} - sortable half year format
(e.g. 2022-H1)
\item \strong{quarter} - \code{jnry_quarter} - sortable quarter format
(e.g. 2022-Q2)
\item \strong{month} - \code{jnry_month} - ISO calendar month basic format
(e.g. 2022-05)
\item \strong{week} - \code{jnry_week} - ISO calendar week extended format
(e.g. 2022-W25)
\item \strong{day} - \code{jnry_day} - ISO calendar date extended format
(e.g. 2022-05-03)
\item \strong{\emph{adhoc}} - \code{jnry_adhoctimetime} - format taken
from user input
}
}

\subsection{Filling time series}{

If \code{fill_time_series = FALSE}, any year, month, etc. not represented in
the input data will not appear in the return data.

If \code{fill_time_series = TRUE}, \code{aggregate_time} will fill in
missing time points with NAs for all variables not measuring time or
included in \code{grouping_vars}. Time series are filled with respect to
grouping variables, spanning the first and last observations within each
group and not globally over the input data.

\code{fill_time_series} is ignored if \code{adhoc_granularity} is provided.
}
}
\description{
Aggregate time granularity of a data set, i.e. take a data set from more to
less granular time scales. \code{aggregate_time} can aggregate data to a set
of standard time granularities (e.g. week, month, year), or it can accept
user-supplied ad hoc intervals (e.g. election periods).
}
\examples{

}
