% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lut.R
\name{build_similarity_dict}
\alias{build_similarity_dict}
\title{Build a similarity dictionary for tokens.}
\usage{
build_similarity_dict(.fml, .base, .target, .method, .threshold)
}
\arguments{
\item{.fml}{A formula defining the preprocessing steps to be applied to the tokens.}

\item{.base}{A data table containing the base dataset.}

\item{.target}{A data table containing the target dataset.}

\item{.method}{A string indicating the method to be used for computing string distance (e.g., "jw" for Jaro-Winkler).}

\item{.threshold}{A numeric value indicating the maximum allowable distance for two tokens to be considered similar.}
}
\value{
Returns a data table with tokens and their corresponding similarity groups.
}
\description{
This function creates a similarity dictionary by comparing tokens from two datasets based on a specified method
and threshold. It returns a data table with tokens and their respective groups based on similarity.
}
\examples{
build_similarity_dict(
  .fml = Nachname ~ normalize_text + word_tokens(.min_length = 3),
  .base = base_table,
  .target = target_table,
  .method = "jw",
  .threshold = 0.05
)
}
