% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{cluster_size_plot}
\alias{cluster_size_plot}
\title{Bar chart of cluster-size distribution (duplicates only)}
\usage{
cluster_size_plot(x, ...)
}
\arguments{
\item{x}{A \code{Match_Overview} object from \code{\link[=summarise_matches]{summarise_matches()}} with
\code{match_type == "duplicates"}.}

\item{...}{Passed to \code{\link[tinyplot:tinyplot]{tinyplot::tinyplot()}}.}
}
\value{
Invisibly, the plotted \code{data.table} (cluster_dist).
}
\description{
Bar chart of cluster-size distribution (duplicates only)
}
