% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{coverage_plot}
\alias{coverage_plot}
\title{Bar chart of match coverage (base and/or target)}
\usage{
coverage_plot(x, ...)
}
\arguments{
\item{x}{A \code{Match_Overview} object from \code{\link[=summarise_matches]{summarise_matches()}}.}

\item{...}{Passed to \code{\link[tinyplot:tinyplot]{tinyplot::tinyplot()}}.}
}
\value{
Invisibly, the plotted \code{data.table}.
}
\description{
Bar chart of match coverage (base and/or target)
}
