% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disamb_party.R
\name{disamb_party}
\alias{disamb_party}
\title{Disambiguate political party names}
\usage{
disamb_party(x, party_ref = y, country = NULL, year = NULL, origin = NULL)
}
\arguments{
\item{origin}{}
}
\value{
A tibble, with a new column for Party Facts id
}
\description{
Disambiguate political party names
}
\examples{

}
