% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparer_tokens.R
\name{fuzzy_tokens}
\alias{fuzzy_tokens}
\title{Fuzzy tokens using igraph components (fast, sparse)}
\usage{
fuzzy_tokens(x, max_dist = 2, method = "osa", min_nchar = 1)
}
\arguments{
\item{x}{Character vector}

\item{max_dist}{Maximum string distance to consider an edge}

\item{method}{stringdist method ("osa", "lv", "jw", ...)}

\item{min_nchar}{Minimum token size}
}
\value{
List of fuzzy tokens (list-column)
}
\description{
Fuzzy tokens using igraph components (fast, sparse)
}
