% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_preparers.R
\name{preapare_search_data}
\alias{preapare_search_data}
\title{Prepare Search Data}
\usage{
preapare_search_data(.preparers, .df, .key)
}
\arguments{
\item{.preparers}{A list of preparer functions created by `search_preparers`.}

\item{.df}{A data frame containing the data to be prepared for search.}

\item{.key}{A character string specifying the key column in the data frame.}
}
\value{
A data.table with prepared search data, including tokens and token counts for each specified column.
}
\description{
This function applies the preparer functions to the specified columns in the input data frame to create a search table.
}
\examples{
search_table_base <- preapare_search_data(prepaers, base_table, "key_base")
}
