% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_preparers.R
\name{search_preparers}
\alias{search_preparers}
\title{Create a list of preparer functions based on a formula syntax.}
\usage{
search_preparers(...)
}
\arguments{
\item{...}{Formula(s) specifying the preparers.}
}
\value{
A list of preparer functions for each specified column.
}
\description{
This function accepts a formula where the left-hand side specifies the column name,
and the right-hand side specifies the sequence of functions to be applied.
}
\examples{
search_preparers(
  Nachname ~ normalize_text + word_tokens,
  Vorname ~ normalize_text + generate_ngrams(3)
)
}
