% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparers_datatable.R
\name{token_shapes}
\alias{token_shapes}
\title{Convert tokens to shape signatures (letter/digit patterns)}
\usage{
token_shapes(tokens)
}
\arguments{
\item{tokens}{A list of character vectors.}
}
\value{
A list of shape tokens.
}
\description{
"MULLER" -> "AAAAAA"
"A12B"   -> "ANNA"
}
