% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_preparers.R
\name{as_metaphone}
\alias{as_metaphone}
\title{Convert Text to Metaphone Encoding}
\usage{
as_metaphone(.text)
}
\arguments{
\item{.text}{A character string or vector to be converted to Metaphone encoding.}
}
\value{
Returns the Metaphone encoded version of the input text.
}
\description{
This function converts a text string to its Metaphone encoding. The Metaphone algorithm is used to
encode words phonetically by reducing them to a simplified representation based on their pronunciation.
}
\examples{
as_metaphone("Café")
as_metaphone("Straße")
}
