% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_preparers.R
\name{as_soundex}
\alias{as_soundex}
\title{Convert Text to Soundex Encoding}
\usage{
as_soundex(.text)
}
\arguments{
\item{.text}{A character string or vector to be converted to Soundex encoding.}
}
\value{
Returns the Soundex encoded version of the input text.
}
\description{
This function converts a text string to its Soundex encoding. The Soundex algorithm is used to
encode words phonetically by reducing them to a simplified representation based on their pronunciation.
}
\examples{
as_soundex("Café")
as_soundex("Straße")
}
