% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_functions.R
\name{deduplicate_table}
\alias{deduplicate_table}
\title{Deduplicate Records}
\usage{
deduplicate_table(.table, .duplicates, .key)
}
\arguments{
\item{.table}{A data table containing the records to be deduplicated.}

\item{.duplicates}{A data table containing the list of likely duplicates.}

\item{.key}{A character string specifying the key column in the data table.}
}
\value{
A data table with duplicates removed.
}
\description{
This function takes a data table and a list of likely duplicates and removes duplicates, keeping only the first occurrence.
}
\examples{
deduplicated_table <- deduplicate(base_table, likely_duplicates, "key_base")
}
