% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disamb_party.R
\name{disamb_party}
\alias{disamb_party}
\title{Disambiguate political party names + match to Party Facts ID}
\usage{
disamb_party(x, party_ref, country, year = NULL, origin = NULL)
}
\arguments{
\item{x}{Data frame in which each observation relates to one political party}

\item{party_ref}{A string identifying the variable in the data, \code{x}, containing political party names or other party identifiers}

\item{country}{A string identifying the country, if single-country data, or vector of strings identifying the country of each observation if multi-country data.}

\item{year}{A four-digit numeric identifying the year of the data, or a string identifying the variable in \code{x} containing the year variable (four-digit numeric). If \code{year = NULL}, assumes the current calendar year.}

\item{origin}{A string identifying the origin format of country names (see \code{countrycode} package). If \code{origin = NULL}, attempts to resolve the origin format automatically.}
}
\value{
A tibble, \code{x}, with new columns for Party Facts id and information on the matching process
}
\description{
Disambiguate political party names + match to Party Facts ID
}
\examples{


}
