% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparers.R
\name{extract_initials}
\alias{extract_initials}
\title{Extract initials from tokens}
\usage{
extract_initials(tokens)
}
\arguments{
\item{tokens}{A list of character vectors.}
}
\value{
A list of character vectors of initials.
}
\description{
Converts tokens to their first-letter initial ("ANNA" -> "A").
}
