% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_preparers.R
\name{generate_ngrams}
\alias{generate_ngrams}
\title{Generate n-grams from text}
\usage{
generate_ngrams(.text, .n)
}
\arguments{
\item{.text}{A character string or vector from which to generate n-grams.}

\item{.n}{An integer specifying the length of each n-gram.}
}
\value{
Returns a list of n-grams generated from the input text. If the text length is less than n, returns
        an empty character vector.
}
\description{
This function generates n-grams from a given text string. An n-gram is a contiguous sequence of n items
from a given sample of text or speech. This function will return a list of all possible n-grams of length n.
}
\examples{
generate_ngrams("hello", 2)
generate_ngrams("an example", 3)
}
