% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinery-package.R
\docType{package}
\name{joinery-package}
\alias{joinery}
\alias{joinery-package}
\title{joinery: Heuristic Index-Based Record Linkage in R}
\description{
Index-based heuristic record linkage for R.
}
\author{
\strong{Maintainer}: Eduard Brüll \email{eduard.bruell@zew.de}

}
\keyword{internal}
