% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinery-package.R
\docType{package}
\name{joinery-package}
\alias{joinery}
\alias{joinery-package}
\title{joinery: Heuristic Index-Based Record Linkage in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Index-based heuristic record linkage for R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://edubruell.github.io/joinery/}
  \item \url{https://github.com/edubruell/joinery}
  \item Report bugs at \url{https://github.com/edubruell/joinery/issues}
}

}
\author{
\strong{Maintainer}: Eduard Brüll \email{eduard.bruell@zew.de}

}
\keyword{internal}
