% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_preparers.R
\name{normalize_text}
\alias{normalize_text}
\title{Normalize text string}
\usage{
normalize_text(.text, .transliteration = "De-ASCII")
}
\arguments{
\item{.text}{A character string or vector to be normalized.}

\item{.transliteration}{A character string specifying the transliteration scheme to be used,
defaulting to "De-ASCII".}
}
\value{
Returns a normalized, upper-case version of the input text, with non-alphanumeric characters
        and extra spaces removed.
}
\description{
This function converts a text string to upper case, transliterates it based on the specified
transliteration scheme, retains only alphanumeric characters and spaces, and removes extra spaces.
}
\examples{
normalize_text("Café Coñac")
normalize_text("Straße", .transliteration = "Latin-ASCII")
}
