% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{prepare_search_data}
\alias{prepare_search_data}
\title{Prepare Data for Record Linkage Search}
\usage{
prepare_search_data(data, id, strategy, ...)
}
\arguments{
\item{data}{A data.frame / tibble / data.table (or db table in other backends).}

\item{id}{Character scalar naming the ID column in `data`.}

\item{strategy}{A `Search_Strategy` object.}

\item{...}{Additional arguments passed to backend-specific methods.}
}
\description{
Prepare Data for Record Linkage Search
}
