% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_preparers.R
\name{use_dictionary}
\alias{use_dictionary}
\title{Use similarity dictionary to group similar tokens together}
\usage{
use_dictionary(.text, .dict)
}
\arguments{
\item{.text}{A character string or vector representing the token to be looked up.}

\item{.dict}{A data table containing the similarity dictionary with tokens and their respective groups.}
}
\value{
Returns the token group corresponding to the input token.
}
\description{
This function looks up a token in the similarity dictionary and returns the corresponding token group for a token.
}
\examples{
dict <- data.table(tokens = c("example", "sample"), token_group = c("example/sample", "example/sample"))
use_dictionary("example", dict)
use_dictionary("nonexistent", dict)
}
